requirements.txt
torch>=2.0.0
transformers>=4.57.0
accelerate>=0.25.0
qwen-vl-utils
flash-attn>=2.5.0
opencv-python>=4.8.0
Pillow>=10.0.0
pynvml>=11.5.0
matplotlib>=3.7.0
numpy>=1.24.0
tqdm>=4.65.0
pyyaml>=6.0
requests>=2.31.0
